test(b): migrate discover tests -> e2e/test_03_discover (#40)
Thirty tests relocated from three sources:
- tests/commands/test_discover.py (14)
- tests/commands/test_discover_multi_source.py (14)
- tests/test_explicit_name_flag.py (2 of 7; the other 5 are source-class
  unit tests that stay for Wave D)

stub_viewer_serve marker preserved (now resolves in tests/e2e/conftest.py
per Task B0). Renamed-config tests use cli(config=<other_path>) per the
new cli kwarg. Postgres-gated class keeps its skipif + db setup as-is.
Single-member classes dropped per Rule M2; Postgres and rename classes
retained because they share fixtures.
